fix: show right vehicles in right place
