+++ app/component/RouteMapContainer.js
@@ -10,6 +10,7 @@ import VehicleMarkerContainer from './map/VehicleMarkerContainer';
 import { getStartTime } from '../util/timeUtils';
 import withBreakpoint from '../util/withBreakpoint';
 import BackButton from './BackButton';
+import { isActiveDate } from '../util/patternUtils';
 
 class RouteMapContainer extends React.PureComponent {
   static propTypes = {
@@ -54,16 +55,18 @@ class RouteMapContainer extends React.PureComponent {
 
     let tripStart;
 
-    const leafletObjs = [
-      <RouteLine key="line" pattern={pattern} />,
-      <VehicleMarkerContainer
-        key="vehicles"
-        direction={pattern.directionId}
-        pattern={pattern.code}
-        headsign={pattern.headsign}
-        tripStart={tripStart}
-      />,
-    ];
+    const leafletObjs = [<RouteLine key="line" pattern={pattern} />];
+    if (isActiveDate(pattern)) {
+      leafletObjs.push(
+        <VehicleMarkerContainer
+          key="vehicles"
+          direction={pattern.directionId}
+          pattern={pattern.code}
+          headsign={pattern.headsign}
+          tripStart={tripStart}
+        />,
+      );
+    }
 
     const showScale = breakpoint === 'large';
 
@@ -158,6 +161,9 @@ export default createFragmentContainer(RouteMapContainerWithVehicles, {
         name
         gtfsId
         ...StopCardHeaderContainer_stop
+      }
+      activeDates: trips {
+        day: activeDates
       }
       ...RouteLine_pattern
     }
